+++ src/services/channels/index.ts
@@ -3,9 +3,7 @@ import { Router } from 'express'
 import { deleteByIdChannelsRouter } from './:channelId/delete'
 import { putByIdChannelsRouter } from './:channelId/put'
 
-const channelsRouter = Router()
+export const channelsRouter = Router()
 
 channelsRouter.use('/', deleteByIdChannelsRouter)
 channelsRouter.use('/', putByIdChannelsRouter)
-
-export { channelsRouter }
