+++ src/services/guilds/index.ts
@@ -9,7 +9,7 @@ import { putByIdGuildsRouter } from './:guildId/put'
 import { getChannelsRouter } from './:guildId/channels/get'
 import { postChannelsRouter } from './:guildId/channels/post'
 
-const guildsRouter = Router()
+export const guildsRouter = Router()
 
 guildsRouter.use('/', postGuildsRouter)
 guildsRouter.use('/', getGuildsRouter)
@@ -19,5 +19,3 @@ guildsRouter.use('/', deleteByIdGuildsRouter)
 guildsRouter.use('/', putByIdGuildsRouter)
 guildsRouter.use('/', getChannelsRouter)
 guildsRouter.use('/', postChannelsRouter)
-
-export { guildsRouter }
