+++ src/services/users/current/index.ts
@@ -1,0 +1,19 @@
+import { Router } from 'express'
+
+import { getCurrentRouter } from './get'
+import { putCurrentRouter } from './put'
+
+export const currentRouter = Router()
+
+export const errorsMessages = {
+  email: {
+    mustBeValid: 'Email must be valid',
+    alreadyConnected: 'You are already connected with this email address'
+  },
+  name: {
+    alreadyConnected: 'You are already connected with this name'
+  }
+}
+
+currentRouter.use('/', getCurrentRouter)
+currentRouter.use('/', putCurrentRouter)
