+++ src/services/users/index.ts
@@ -6,7 +6,7 @@ import { discordRouter } from './oauth2/discord'
 import { githubRouter } from './oauth2/github'
 import { googleRouter } from './oauth2/google'
 import { refreshTokenRouter } from './refreshToken'
-import { resetPasswordRouter } from './resetPassword'
+import { resetPasswordRouter } from './reset-password'
 import { signinRouter } from './signin'
 import { signoutRouter } from './signout'
 import { signupRouter } from './signup'
